backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
tensorflow
numpy
google-genai
pydantic
